+++ src/script/mct/settings/cbac.lua
@@ -19,16 +19,16 @@ option_cbac_army_limit_player:set_default_value(10500);
 
 local option_cbac_army_limit_ai = mct_mod:add_new_option("army_limit_ai", "slider");
 option_cbac_army_limit_ai:set_text("AI Limit (Base)");
-option_cbac_army_limit_ai:set_tooltip_text("How many points per AI army?\n\n[[col:yellow]]Warning! A smaller value will
-                                           cause AI turns to take longer.[[/col]]");
+option_cbac_army_limit_ai:set_tooltip_text("How many points per AI army?\n\n[[col:yellow]]Warning! A smaller value " ..
+                                           "will cause AI turns to take longer.[[/col]]");
 option_cbac_army_limit_ai:slider_set_min_max(7000, 40000);
 option_cbac_army_limit_ai:slider_set_step_size(500);
 option_cbac_army_limit_ai:set_default_value(12000);
 
 local option_cbac_hero_cap = mct_mod:add_new_option("hero_cap", "slider");
 option_cbac_hero_cap:set_text("Allowed Heroes per Army");
-option_cbac_hero_cap:set_tooltip_text("Exceeding this number of embedded heroes will trigger the same penalties as being
-                                      over the cost limit.");
+option_cbac_hero_cap:set_tooltip_text("Exceeding this number of embedded heroes will trigger the same penalties as " ..
+                                      "being over the cost limit.");
 option_cbac_hero_cap:slider_set_min_max(0, 19);
 option_cbac_hero_cap:slider_set_step_size(1);
 option_cbac_hero_cap:set_default_value(2);
@@ -49,25 +49,25 @@ option_cbac_limit_rank:set_default_value(2);
 
 local option_cbac_limit_step = mct_mod:add_new_option("limit_step", "slider");
 option_cbac_limit_step:set_text("Dynamic Limit increase step size");
-option_cbac_limit_step:set_tooltip_text("Every time a lord's level increase triggers a limit increase, it goes up by
-                                        this amount (if limit is set to dynamic);.");
+option_cbac_limit_step:set_tooltip_text("Every time a lord's level increase triggers a limit increase, it goes up by" ..
+                                        " this amount (if limit is set to dynamic);.");
 option_cbac_limit_step:slider_set_min_max(100, 5000);
 option_cbac_limit_step:slider_set_step_size(100);
 option_cbac_limit_step:set_default_value(1000);
 
 local option_cbac_limit_deceleration = mct_mod:add_new_option("limit_deceleration", "slider");
 option_cbac_limit_deceleration:set_text("Dynamic Limit deceleration");
-option_cbac_limit_deceleration:set_tooltip_text("At every step, the limit increase slows down by this amount, so that
-                                                the limit rises fast at first, but the increase slows down at higher
-                                                levels.");
+option_cbac_limit_deceleration:set_tooltip_text("At every step, the limit increase slows down by this amount, so " ..
+                                                "that the limit rises fast at first, but the increase slows down " ..
+                                                "at higher levels.");
 option_cbac_limit_deceleration:slider_set_min_max(0, 2000);
 option_cbac_limit_deceleration:slider_set_step_size(10);
 option_cbac_limit_deceleration:set_default_value(50);
 
 local option_cbac_supply_lines = mct_mod:add_new_option("supply_lines", "checkbox");
 option_cbac_supply_lines:set_text("Enable Cost-based Supply Lines");
-option_cbac_supply_lines:set_tooltip_text("Adds Supply Lines unit upkeep increases based on army cost. Use the submod
-                                          to disable the normal Supply Lines!");
+option_cbac_supply_lines:set_tooltip_text("Adds Supply Lines unit upkeep increases based on army cost. Use the " ..
+                                          "submod to disable the normal Supply Lines!");
 option_cbac_supply_lines:set_default_value(false);
 
 local option_cbac_army_limit_fl_bonus = mct_mod:add_new_option("army_limit_fl_bonus", "slider");
@@ -81,8 +81,8 @@ local mct_section_ai = mct_mod:add_new_section("3-cbac-ai", "AI Behaviour Option
 
 local option_cbac_upgrade_ai = mct_mod:add_new_option("upgrade_ai_armies", "checkbox");
 option_cbac_upgrade_ai:set_text("Enable Upgraded AI Armies");
-option_cbac_upgrade_ai:set_tooltip_text("Brings AI armies closer to their army limit if they have the gold to afford
-                                        it.");
+option_cbac_upgrade_ai:set_tooltip_text("Brings AI armies closer to their army limit if they have the gold to afford" ..
+                                        " it.");
 option_cbac_upgrade_ai:set_default_value(false);
 
 local option_cbac_upgrade_grace_period = mct_mod:add_new_option("upgrade_grace_period", "slider");
@@ -94,8 +94,8 @@ option_cbac_upgrade_grace_period:set_default_value(20);
 
 local option_cbac_autolevel_ai = mct_mod:add_new_option("auto_level_ai_lords", "slider");
 option_cbac_autolevel_ai:set_text("Auto-level AI lords");
-option_cbac_autolevel_ai:set_tooltip_text("AI lords will have a minimum level based on the turn number. 0=off 1=slower
-                                          increase 5=faster increase");
+option_cbac_autolevel_ai:set_tooltip_text("AI lords will have a minimum level based on the turn number. " ..
+                                          "0=off 1=slower increase 5=faster increase");
 option_cbac_autolevel_ai:slider_set_min_max(0, 5);
 option_cbac_autolevel_ai:slider_set_step_size(1);
 option_cbac_autolevel_ai:set_default_value(3);
@@ -104,7 +104,7 @@ local mct_section_sl = mct_mod:add_new_section("4-cbac-sl", "Setting Lock", fals
 
 local option_cbac_settings_locked = mct_mod:add_new_option("settings_locked", "checkbox");
 option_cbac_settings_locked:set_text("Lock settings during campaigns");
-option_cbac_settings_locked:set_tooltip_text("If enabled, you can't change these settings during a campaign, only from
-                                             the main menu.");
+option_cbac_settings_locked:set_tooltip_text("If enabled, you can't change these settings during a campaign, only " ..
+                                             "from the main menu.");
 option_cbac_settings_locked:set_default_value(false);
 option_cbac_settings_locked:set_read_only(true);