+++ src/script/_lib/mod/cbac.lua
@@ -64,8 +64,11 @@ end
 -- GENERIC --
 
 function cbac:log(str)
-  if config["logging_enabled"] then
-    out('CBAC ' .. str);
+  if cbac:get_config("logging_enabled") then
+    local log_file = io.open("wolfy_mods_log.txt","a");
+    log_file:write("\n[cbac] " .. str);
+    log_file:flush();
+    log_file:close();
   end
 end
 
@@ -78,9 +81,8 @@ function cbac:get_config(config_key)
     local mct = get_mct();
 
     if mct ~= nil then
-      cbac:log("Loading config from MCT: " .. config_key);
       local mod_cfg = mct:get_mod_by_key("wolfy_cost_based_army_caps");
-      config[config_key] = mod_cfg:get_option_by_key(config_key):get_finalized_setting();
+      return mod_cfg:get_option_by_key(config_key):get_finalized_setting();
     end
   end
 
@@ -152,6 +154,14 @@ function cbac:get_unit_cost(unit)
     return 0;
   else
     return unit:get_unit_custom_battle_cost();
+  end
+end
+
+function cbac:get_unit_cost_from_key(unit_key)
+  if _is_free_unit(unit_key) then
+    return 0;
+  else
+    return cco("CcoMainUnitRecord", unit_key):Call("Cost");
   end
 end
 
@@ -247,11 +257,11 @@ function cbac:get_army_queued_units_cost()  -- TODO CHECK NURGLE AND ROR
         unit_card:SimulateMouseOn();
 
         local ok, err = pcall(function()
-          local unit_info = find_uicomponent(core:get_ui_root(), "hud_campaign", "unit_information_parent", "unit_info_panel_holder_parent", "unit_info_panel_holder");
-          local unit_context_id = unit_info:GetContextObjectId("CcoUnitDetails");
-          local unit_key_head = string.gsub(unit_context_id, "RecruitmentUnit_", "");
-          local unit_key = string.gsub(unit_key_head, "_%d+_%d+_%d+_%d+$", "");
-          local unit_cost = cco("CcoMainUnitRecord", unit_key):Call("Cost");
+          local unit_info = find_uicomponent(core:get_ui_root(), "hud_campaign", "unit_information_parent",
+                                             "unit_info_panel_holder_parent", "unit_info_panel_holder");
+          local unit_key = string.gsub(string.gsub(unit_info:GetContextObjectId("CcoUnitDetails"),
+                                                   "RecruitmentUnit_", ""), "_%d+_%d+_%d+_%d+$", "");
+          local unit_cost = cbac:get_unit_cost_from_key(unit_key);
 
           cbac:log("Value of " .. unit_key .. ": " .. unit_cost);
           queued_units_cost = queued_units_cost + unit_cost;